+++ Source/Actors/Enemies/Golem2/Golem2.cpp
@@ -1,10 +1,9 @@
-#include "Ghost.h"
-#include "../../Game.h"
-#include "../../Components/PhysicsComponents/RigidBodyComponent.h"
-#include "../../Components/DrawComponents/DrawAnimatedComponent.h"
-#include "../../Components/ColliderComponents/AABBColliderComponent.h"
-#include "../Items/Collectible/CollectibleItem.h"
-#include "../Projectile/GhostAttack/GhostAttack.h"
+#include "Golem2.h"
+#include "../../../Game.h"
+#include "../../../Components/PhysicsComponents/RigidBodyComponent.h"
+#include "../../../Components/DrawComponents/DrawAnimatedComponent.h"
+#include "../../../Components/ColliderComponents/AABBColliderComponent.h"
+#include "../../Items/Collectible/CollectibleItem.h"
 
 #include <random>
 
@@ -13,22 +12,22 @@ const std::string WALKING_ANIMATION = "walking";
 const std::string ATTACK_ANIMATION = "attack";
 
 
-Ghost::Ghost(Game *game, Vector2 position) : Enemy(game) {
+Golem2::Golem2(Game *game, Vector2 position) : Enemy(game) {
 
     SetPosition(position);
 
-    mMaxHealth = 100.0f;
+    mMaxHealth = 130.0f;
     mCurrentHealth = mMaxHealth;
     mIsDead = false;
 
-    mWalkSpeed =  60.0f;
+    mWalkSpeed =  70.0f;
 
     mRigidBodyComponent = new RigidBodyComponent(this, 1.0f, 5.0f);
-    int GhostSize = 99;
+    int golem2Size = 99;
     int colliderSize = 80;
     
-    int offsetX = (GhostSize - colliderSize) / 2;
-    int offsetY = (GhostSize - colliderSize) / 2;
+    int offsetX = (golem2Size - colliderSize) / 2;
+    int offsetY = (golem2Size - colliderSize) / 2;
     
     mColliderComponent = new AABBColliderComponent(
         this, 
@@ -41,26 +40,26 @@ Ghost::Ghost(Game *game, Vector2 position) : Enemy(game) {
         true
     );
 
-    mDamageAttack = 15.0f;
+    mDamageAttack = 5.0f;
 
-    mAttackCooldown = 2.0f;
+    mAttackCooldown = 1.5f;
 
     mDrawComponent = new DrawAnimatedComponent(
         this,
-        "../Assets/Sprites/Ghost/ghost.png",
-        "../Assets/Sprites/Ghost/ghost.json",
+        "../Assets/Sprites/Golem_2/golem2.png",
+        "../Assets/Sprites/Golem_2/golem2.json",
         100
     );
 
-    mDrawComponent->AddAnimation(IDLE_ANIMATION, {0});
-    mDrawComponent->AddAnimation(ATTACK_ANIMATION, {4,6,9, 2});
-    mDrawComponent->AddAnimation(WALKING_ANIMATION, {10,1,7,8});
+    mDrawComponent->AddAnimation(IDLE_ANIMATION, {1});
+    mDrawComponent->AddAnimation(ATTACK_ANIMATION, {3, 4, 5});
+    mDrawComponent->AddAnimation(WALKING_ANIMATION, {1, 3, 0, 9});
     mDrawComponent->SetAnimFPS(10.f);
 
     mDrawComponent->SetAnimation(IDLE_ANIMATION);
 }
 
-void Ghost::Attack() {
+void Golem2::Attack() {
     if (mAttackCooldown >= 0.0f) {
         return;
     }
@@ -77,17 +76,13 @@ void Ghost::Attack() {
     if (mGame->GetAudioSystem()->GetSoundState(mAttackSound) != SoundState::Playing)
         mAttackSound = mGame->GetAudioSystem()->PlaySound("monster_attack.wav", false);
 
-    Vector2 playerPos = mGame->GetPlayer()->GetPosition();
-    Vector2 direction = playerPos - GetPosition();
-    direction.Normalize();
-
-    new GhostAttack(mGame, GetPosition(), direction * 200.0f, mDamageAttack); 
+    player->TakeDamage(mDamageAttack);
     ManageAnimations();
 
-    mAttackCooldown = 2.0f; // reset
+    mAttackCooldown = 0.8f; // reset
 }
 
-void Ghost::OnUpdate(float deltaTime) {
+void Golem2::OnUpdate(float deltaTime) {
 
     //DebugColliderPosition();
 
@@ -116,7 +111,7 @@ void Ghost::OnUpdate(float deltaTime) {
 
     if (distance > 0.0f) toPlayer *= 1/distance; // normalize
 
-    if (distance < 500.0f) {
+    if (distance < 60.0f) {
         Attack();
     } else {
         mIsWalking = true;
@@ -130,7 +125,7 @@ void Ghost::OnUpdate(float deltaTime) {
     ManageAnimations();
 }
 
-void Ghost::Kill() {
+void Golem2::Kill() {
     if (mState == ActorState::Destroy) {
         return; // ensure Kill is executed only once
     }
@@ -160,7 +155,7 @@ void Ghost::Kill() {
     // TODO: Add more items to drop
 }
 
-void Ghost::ManageAnimations() const {
+void Golem2::ManageAnimations() const {
     if (mIsAttacking) {
         mDrawComponent->SetAnimation(ATTACK_ANIMATION);
     } else if (!mIsWalking) {
@@ -170,7 +165,7 @@ void Ghost::ManageAnimations() const {
     }
 }
 
-void Ghost::OnCollision(float minOverlap, AABBColliderComponent *other) {
+void Golem2::OnCollision(float minOverlap, AABBColliderComponent *other) {
 
     if (other->GetLayer() == ColliderLayer::Player) {
         // get the player damage
